+++ CHANGELOG.md
@@ -3,6 +3,9 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 
+- Set default config namespace for Get/Try/Require methods in Go SDK.
+  [4802](https://github.com/pulumi/pulumi/pull/4802)
+
 - Improve typing for Go SDK secret config values
  [#4800](https://github.com/pulumi/pulumi/pull/4800)